Do not store computed sweep frequency on trigger. Only check for overflow.
It fixes *at least* background ambiance sounds in Metroid II and rotating sfx in Tetris. I am *so* having a drink to celebrate this!

Turns out "frequency calculation and overflow check are performed immediately" only means that the calculated frequency is used for the overflow check. It does *not* mean it should be immediately stored back to NR13 and NR14. That was *not* clear to me. x_x